Remove Cloudflare optimization cookie feature (Mirage/Polish/Fonts)
The performance module advertised active Cloudflare optimizations by
setting the nfd-enable-cf-opt cookie. Adoption of the toggles is
negligible, Mirage is deprecated, and the remaining optimizations are
marginal versus page caching while the cookie risked making responses
uncacheable under cookieless (load-test/bot) traffic.

Remove the feature entirely: the cookie mechanism, the Fonts subsystem
and its UI section, and the Polish/Mirage toggles in Image Optimization.
Cloudflare Polish/Fonts will be enabled at the zone level instead.

Keep the Newfold image-CDN optimizer (auto/bulk/WebP/lazy-load). Retain
a one-time .htaccess cleanup (LegacyCookieCleanup) so installs upgrading
from before 3.8.0 still get the stale Set-Cookie block stripped.
