--- src/sections/FontOptimization/stylesheet.css
@@ -1,4 +1,0 @@
-.newfold-font-optimization .nfd-button--upsell {
-	position: absolute;
-	top: 0;
-}
