+++ src/styles/styles.css
@@ -3,7 +3,6 @@
 @import "../sections/LinkPrefetch/stylesheet.css";
 @import "../sections/JetpackBoost/stylesheet.css";
 @import "../sections/ImageOptimization/stylesheet.css";
-@import "../sections/FontOptimization/stylesheet.css";
 @import "./branding.css";
 @import "../components/App/stylesheet.css";
